extracted scorecard into its own class
